1.20.1 Port

+++ common/src/main/java/de/mrjulsen/trafficcraft/block/AsphaltSlope.java
@@ -76,7 +76,7 @@ public class AsphaltSlope extends AsphaltBlock implements SimpleWaterloggedBlock
             level.levelEvent(player, LevelEvent.PARTICLES_DESTROY_BLOCK, pos, Block.getId(state));
         } else {
             if (!(player.isCreative() || player.isSpectator())) {
-                dropResources(state.getBlock().defaultBlockState(), level, pos.offset(0, 1f / 8f * (state.getValue(BlockStateProperties.LAYERS) + 1), 0));
+                dropResources(state.getBlock().defaultBlockState(), level, pos.offset(0, (int)(1f / 8f * (state.getValue(BlockStateProperties.LAYERS) + 1)), 0));
             }
             tool.hurtAndBreak(1, player, (p) -> {
                 player.broadcastBreakEvent(player.getItemInHand(InteractionHand.MAIN_HAND) == tool ? EquipmentSlot.MAINHAND : EquipmentSlot.OFFHAND);
